Remove pkg/healthcheck
This original came from the proxy plugin, but as that is now external
this can be removed.
However the k8s plugin also used it. That code has been removed in a
recent PR; meaning this code has no user anymore and can be included
in the proxy repo and removed from this.

Signed-off-by: Miek Gieben <miek@miek.nl>